+++ .recon.config.mjs
@@ -2,13 +2,14 @@ export default {
   commands: {
     all: {
       gather: {
-        files: ["./*"],
+        files: ["./"],
       },
     },
     code: {
       gather: {
         notes: "IMPORTANT: Only return the code, no commentary. Do not wrap the code in backticks.",
+        files: ["./"],
       }
-    }
+    },
   },
 };